+++ lib/src/http.dart
@@ -1,9 +1,18 @@
+import 'dart:async';
+import 'dart:io' as io;
+
 import 'package:dio/dio.dart';
+import 'package:dio/io.dart';
+import 'package:dio_http2_adapter/dio_http2_adapter.dart';
+import 'package:flutter/foundation.dart';
 
 import 'coverters/converter.dart';
+import 'http2/connection_manager_impl.dart';
 import 'intercaptors/log_interceptor.dart';
-import 'options.dart';
+import 'options/options.dart';
 import 'request.dart';
+import 'system_proxy.dart';
+import 'utils/utils.dart';
 
 part 'session.dart';
 
@@ -18,7 +27,7 @@ typedef Parameters = Map<String, dynamic>;
 
 /// 请求体
 enum ContentType {
-  /// 原始
+  /// text
   raw,
 
   /// json
@@ -34,9 +43,9 @@ enum ContentType {
   String get value {
     switch (this) {
       case ContentType.raw:
-        return Headers.textPlainContentType;
+        return io.ContentType.text.toString();
       case ContentType.json:
-        return Headers.jsonContentType;
+        return io.ContentType.json.toString();
       case ContentType.urlencoded:
         return Headers.formUrlEncodedContentType;
       case ContentType.multipart:
@@ -46,15 +55,18 @@ enum ContentType {
 
   /// 从 content-type 值获取 [ContentType]
   static ContentType? tryParse(String? value) {
-    switch (value) {
-      case Headers.textPlainContentType:
-        return ContentType.raw;
-      case Headers.jsonContentType:
-        return ContentType.json;
-      case Headers.formUrlEncodedContentType:
-        return ContentType.urlencoded;
-      case Headers.multipartFormDataContentType:
-        return ContentType.multipart;
+    if (value != null) {
+      final mediaType = io.ContentType.parse(value).mimeType;
+      switch (mediaType) {
+        case Headers.textPlainContentType:
+          return ContentType.raw;
+        case Headers.jsonContentType:
+          return ContentType.json;
+        case Headers.formUrlEncodedContentType:
+          return ContentType.urlencoded;
+        case Headers.multipartFormDataContentType:
+          return ContentType.multipart;
+      }
     }
     return null;
   }
@@ -78,6 +90,9 @@ final class Http {
   late final Dio _dio;
 
   late HttpBaseOptions _options;
+
+  /// 环境变量
+  Map<String, String> environment = {};
 
   /// 配置
   HttpBaseOptions get options => _options;
@@ -106,13 +121,14 @@ final class Http {
   ///
   /// [transformer] 允许在将 请求数据发送至服务器/响应数据从服务器接收 之前进行更改
   /// 这仅适用于具有有效载荷的请求。
-  void config({
+  Future<void> config({
     Dio? dio,
     HttpBaseOptions? options,
     Iterable<Interceptor>? interceptors,
+    HttpClientOptions? httpClientOptions,
     HttpClientAdapter? httpClientAdapter,
     Transformer? transformer,
-  }) {
+  }) async {
     _options = options ?? HttpBaseOptions();
     _dio = dio ?? Dio(_options);
 
@@ -147,6 +163,72 @@ final class Http {
 
     if (httpClientAdapter != null) {
       _dio.httpClientAdapter = httpClientAdapter;
+    } else if (httpClientOptions != null && httpClientOptions.enable) {
+      String? proxyHost;
+      int? proxyPort;
+      final String? proxy = await SystemProxy.getProxy();
+      if (proxy != null && proxy.isNotEmpty) {
+        environment['http_proxy'] = proxy;
+        environment['https_proxy'] = proxy;
+        final List<String> split = proxy.split(':');
+        try {
+          proxyHost = split[0];
+          proxyPort = int.parse(split[1]);
+        } catch (_) {}
+      }
+      if (httpClientOptions.h2) {
+        _dio.httpClientAdapter = Http2Adapter(
+          ConnectionManagerImpl(
+            idleTimeout: 10000,
+            // Ignore bad certificate
+            onClientCreate: (_, config) => config.onBadCertificate = (_) => true,
+            proxyHost: proxyHost,
+            proxyPort: proxyPort,
+          ),
+        );
+      } else if (_dio.httpClientAdapter is IOHttpClientAdapter) {
+        final IOHttpClientAdapter httpClientAdapter = _dio.httpClientAdapter as IOHttpClientAdapter;
+        httpClientAdapter.createHttpClient = () {
+          io.HttpClient client = io.HttpClient()..idleTimeout = const Duration(seconds: 3);
+          if (httpClientOptions.pKCSPath != null) {
+            final io.SecurityContext sc = io.SecurityContext();
+            //file为证书路径
+            sc.setTrustedCertificates(
+              httpClientOptions.pKCSPath!,
+              password: httpClientOptions.pKCSPwd,
+            );
+            client = io.HttpClient(context: sc);
+          }
+          client.badCertificateCallback = (io.X509Certificate cert, String host, int port) => true;
+          client.idleTimeout = const Duration(seconds: 15);
+          if (httpClientOptions.pem != null) {
+            client.badCertificateCallback = (io.X509Certificate cert, String host, int port) =>
+                cert.pem == httpClientOptions.pem;
+          } else {
+            client.badCertificateCallback =
+                (io.X509Certificate cert, String host, int port) => true;
+          }
+          if (!kReleaseMode) {
+            client.findProxy = (Uri url) {
+              /// 因为是异步方法代理设置会有一个请求的延迟;
+              /// 解决办法:
+              ///     1、原生主动通知代理变化
+              ///     2、每次请求前进行获取代理
+              /// 目前每次findProxy时获取，有延迟，因为提供测试使用，暂定
+              SystemProxy.getProxy().then((String? proxy) {
+                if (proxy?.isNotEmpty ?? false) {
+                  environment['http_proxy'] = proxy!;
+                  environment['https_proxy'] = proxy;
+                } else {
+                  environment.clear();
+                }
+              });
+              return io.HttpClient.findProxyFromEnvironment(url, environment: environment);
+            };
+          }
+          return client;
+        };
+      }
     }
 
     if (transformer != null) {
@@ -161,6 +243,7 @@ final class Http {
     Parameters? queryParameters,
     HttpOptions<BaseResp<T>, T>? options,
     CancelToken? cancelToken,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().delete(
       path: path,
@@ -168,6 +251,7 @@ final class Http {
       queryParameters: queryParameters,
       options: options,
       cancelToken: cancelToken,
+      retryOptions: retryOptions,
     );
   }
 
@@ -177,12 +261,14 @@ final class Http {
     Object? data,
     HttpOptions<BaseResp<T>, T>? options,
     CancelToken? cancelToken,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().deleteUri(
       uri: uri,
       data: data,
       options: options,
       cancelToken: cancelToken,
+      retryOptions: retryOptions,
     );
   }
 
@@ -198,6 +284,7 @@ final class Http {
     bool? deleteOnError,
     FileAccessMode? fileAccessMode,
     String? lengthHeader,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<ResponseBody>().download(
       path: path,
@@ -210,6 +297,7 @@ final class Http {
       deleteOnError: deleteOnError,
       fileAccessMode: fileAccessMode,
       lengthHeader: lengthHeader,
+      retryOptions: retryOptions,
     );
   }
 
@@ -224,6 +312,7 @@ final class Http {
     bool? deleteOnError,
     FileAccessMode? fileAccessMode,
     String? lengthHeader,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<ResponseBody>().downloadUri(
       uri: uri,
@@ -235,6 +324,7 @@ final class Http {
       deleteOnError: deleteOnError,
       fileAccessMode: fileAccessMode,
       lengthHeader: lengthHeader,
+      retryOptions: retryOptions,
     );
   }
 
@@ -253,6 +343,7 @@ final class Http {
     bool? deleteOnError,
     FileAccessMode? fileAccessMode,
     String? lengthHeader,
+    HttpRetryOptions? retryOptions,
   }) =>
       session<T>().fetch<E>(
         path: path,
@@ -268,6 +359,7 @@ final class Http {
         deleteOnError: deleteOnError,
         fileAccessMode: fileAccessMode,
         lengthHeader: lengthHeader,
+        retryOptions: retryOptions,
       );
 
   /// get
@@ -278,6 +370,7 @@ final class Http {
     HttpOptions<BaseResp<T>, T>? options,
     CancelToken? cancelToken,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().get(
       path: path,
@@ -286,6 +379,7 @@ final class Http {
       options: options,
       cancelToken: cancelToken,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -296,6 +390,7 @@ final class Http {
     HttpOptions<BaseResp<T>, T>? options,
     CancelToken? cancelToken,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().getUri(
       uri: uri,
@@ -303,6 +398,7 @@ final class Http {
       options: options,
       cancelToken: cancelToken,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -313,6 +409,7 @@ final class Http {
     Parameters? queryParameters,
     HttpOptions<BaseResp<T>, T>? options,
     CancelToken? cancelToken,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().head(
       path: path,
@@ -320,6 +417,7 @@ final class Http {
       queryParameters: queryParameters,
       options: options,
       cancelToken: cancelToken,
+      retryOptions: retryOptions,
     );
   }
 
@@ -329,12 +427,14 @@ final class Http {
     Object? data,
     HttpOptions<BaseResp<T>, T>? options,
     CancelToken? cancelToken,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().headUri(
       uri: uri,
       data: data,
       options: options,
       cancelToken: cancelToken,
+      retryOptions: retryOptions,
     );
   }
 
@@ -347,6 +447,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().patch(
       path: path,
@@ -356,6 +457,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -367,6 +469,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().patchUri(
       uri: uri,
@@ -375,6 +478,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -387,6 +491,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().post(
       path: path,
@@ -396,6 +501,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -407,6 +513,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().postUri(
       uri: uri,
@@ -415,6 +522,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -427,6 +535,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().put(
       path: path,
@@ -436,6 +545,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -447,6 +557,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().putUri(
       uri: uri,
@@ -455,6 +566,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -474,6 +586,7 @@ final class Http {
     bool? deleteOnError,
     FileAccessMode? fileAccessMode,
     String? lengthHeader,
+    HttpRetryOptions? retryOptions,
   }) =>
       session<T>().request(
         path: path,
@@ -490,6 +603,7 @@ final class Http {
         deleteOnError: deleteOnError,
         fileAccessMode: fileAccessMode,
         lengthHeader: lengthHeader,
+        retryOptions: retryOptions,
       );
 
   /// request uri
@@ -507,6 +621,7 @@ final class Http {
     bool? deleteOnError,
     FileAccessMode? fileAccessMode,
     String? lengthHeader,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().requestUri(
       uri: uri,
@@ -522,6 +637,7 @@ final class Http {
       deleteOnError: deleteOnError,
       fileAccessMode: fileAccessMode,
       lengthHeader: lengthHeader,
+      retryOptions: retryOptions,
     );
   }
 
@@ -557,6 +673,7 @@ final class Http {
     dynamic savePath,
     Duration? sendTimeout,
     ValidateStatus? validateStatus,
+    HttpRetryOptions? retryOptions,
   }) =>
       _DefaultSession<T>(
         cancelToken: cancelToken,
@@ -589,6 +706,7 @@ final class Http {
         savePath: savePath,
         sendTimeout: sendTimeout,
         validateStatus: validateStatus,
+        retryOptions: retryOptions,
       );
 
   /// upload
@@ -601,6 +719,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().upload(
       path: path,
@@ -611,6 +730,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 
@@ -623,6 +743,7 @@ final class Http {
     CancelToken? cancelToken,
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
+    HttpRetryOptions? retryOptions,
   }) {
     return session<T>().uploadUri(
       uri: uri,
@@ -632,6 +753,7 @@ final class Http {
       cancelToken: cancelToken,
       onSendProgress: onSendProgress,
       onReceiveProgress: onReceiveProgress,
+      retryOptions: retryOptions,
     );
   }
 }